+++ packages/analytics/src/index.ts
@@ -27,3 +27,4 @@ export * from './seasonality';
 export * from './valuationHistory';
 export * from './fundingAnalytics';
 export * from './bookAnalytics';
+export * from './dexAnalytics';